# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JuneXDeployment API — Discord bot hosting platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deployments:
    get:
      operationId: listDeployments
      tags: [deployments]
      summary: List current user's deployments
      responses:
        "200":
          description: List of deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"
    post:
      operationId: createDeployment
      tags: [deployments]
      summary: Create and deploy a bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeploymentInput"
      responses:
        "201":
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}:
    get:
      operationId: getDeployment
      tags: [deployments]
      summary: Get a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDeployment
      tags: [deployments]
      summary: Delete a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /deployments/{id}/start:
    post:
      operationId: startDeployment
      tags: [deployments]
      summary: Start a stopped deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}/stop:
    post:
      operationId: stopDeployment
      tags: [deployments]
      summary: Stop a running deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}/restart:
    post:
      operationId: restartDeployment
      tags: [deployments]
      summary: Restart a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restarted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}/logs:
    get:
      operationId: getDeploymentLogs
      tags: [deployments]
      summary: Get logs for a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentLogs"
  /deployments/{id}/env:
    patch:
      operationId: updateDeploymentEnv
      tags: [deployments]
      summary: Update environment variables for a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnvVarsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Deployment:
      type: object
      required: [id, userId, templateId, templateName, botName, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        templateId:
          type: integer
        templateName:
          type: string
        botName:
          type: string
        herokuAppId:
          type: ["string", "null"]
        status:
          type: string
          enum: [queued, building, online, offline, error, suspended]
        createdAt:
          type: string
        updatedAt:
          type: string
        templateThumbnail:
          type: ["string", "null"]
        envVars:
          type: object
          additionalProperties:
            type: string
    DeploymentInput:
      type: object
      required: [templateId, botName, envVars]
      properties:
        templateId:
          type: integer
        botName:
          type: string
        envVars:
          type: object
          additionalProperties:
            type: string
    DeploymentLogs:
      type: object
      required: [lines]
      properties:
        lines:
          type: array
          items:
            type: string
    EnvVarsUpdate:
      type: object
      required: [envVars]
      properties:
        envVars:
          type: object
          additionalProperties:
            type: string
...